+++ src/MachinaGrasshopper/Program/Display.cs
@@ -19,31 +19,34 @@ namespace MachinaGrasshopper.Program
         public Display() : base(
             "DisplayProgram",
             "DisplayProgram",
-            "Returns a human-readable representation of a list of Actions.",
+            "Returns a string representation of a robotic program in its native language.",
             "Machina",
             "Program")
         { }
+
         public override GH_Exposure Exposure => GH_Exposure.primary;
-        public override Guid ComponentGuid => new Guid("4f53f8c0-014b-4fd2-b764-71e9e49cf67d");
+        public override Guid ComponentGuid => new Guid("2d842a4d-070c-4a06-925a-ea01b32fd610");
         protected override System.Drawing.Bitmap Icon => Properties.Resources.Program_Display;
 
         protected override void RegisterInputParams(GH_InputParamManager pManager)
         {
-            pManager.AddGenericParameter("Actions", "A", "The list of Actions that conforms a program.", GH_ParamAccess.item);
+            pManager.AddGenericParameter("RobotProgram", "P", "A compiled Robot Program.", GH_ParamAccess.item);
         }
 
         protected override void RegisterOutputParams(GH_OutputParamManager pManager)
         {
-            pManager.AddTextParameter("Program", "P", "Human-readable representation of the program.", GH_ParamAccess.item);
+            pManager.AddTextParameter("Code", "C", "Program code in the robot's native language.", GH_ParamAccess.list);
         }
 
         protected override void SolveInstance(IGH_DataAccess DA)
         {
-            Machina.Action a = null;
+            Machina.Types.Data.RobotProgram program = null;
 
-            if (!DA.GetData(0, ref a)) return;
+            if (!DA.GetData(0, ref program)) return;
+            
+            List<string> code = new List<string>();
 
-            DA.SetData(0, a.ToString());
+            DA.SetDataList(0, program.ToStringList());
         }
     }
 }